[feature/ms-274] update function push_to_report_cashflow in PurchaseOrder

+++ apps/shared/scripts.py
@@ -50,7 +50,7 @@ from ..sales.opportunity.serializers import CommonOpportunityUpdate
 from ..sales.purchasing.models import PurchaseRequestProduct, PurchaseRequest, PurchaseOrderProduct, \
     PurchaseOrderRequestProduct, PurchaseOrder
 from ..sales.quotation.models import QuotationIndicatorConfig, Quotation, QuotationIndicator, QuotationAppConfig
-from ..sales.report.models import ReportRevenue, ReportPipeline, ReportInventorySub
+from ..sales.report.models import ReportRevenue, ReportPipeline, ReportInventorySub, ReportCashflow
 from ..sales.saleorder.models import SaleOrderIndicatorConfig, SaleOrderProduct, SaleOrder, SaleOrderIndicator, \
     SaleOrderAppConfig
 from apps.sales.report.models import ReportRevenue, ReportProduct, ReportCustomer
@@ -1324,14 +1324,21 @@ def update_date_approved():
     print('Done!')
 
 
-def reset_and_run_reports_sale():
-    ReportRevenue.objects.all().delete()
-    ReportCustomer.objects.all().delete()
-    ReportProduct.objects.all().delete()
-    for sale_order in SaleOrder.objects.filter(system_status__in=[2, 3]):
-        SaleOrder.push_to_report_revenue(sale_order)
-        SaleOrder.push_to_report_product(sale_order)
-        SaleOrder.push_to_report_customer(sale_order)
+def reset_and_run_reports_sale(run_type=0):
+    if run_type == 0:  # run report revenue, customer, product
+        ReportRevenue.objects.all().delete()
+        ReportCustomer.objects.all().delete()
+        ReportProduct.objects.all().delete()
+        for sale_order in SaleOrder.objects.filter(system_status__in=[2, 3]):
+            SaleOrder.push_to_report_revenue(sale_order)
+            SaleOrder.push_to_report_product(sale_order)
+            SaleOrder.push_to_report_customer(sale_order)
+    if run_type == 1:  # run report cashflow
+        ReportCashflow.objects.all().delete()
+        for sale_order in SaleOrder.objects.filter(system_status__in=[2, 3]):
+            SaleOrder.push_to_report_cashflow(sale_order)
+        for purchase_order in PurchaseOrder.objects.filter(system_status__in=[2, 3]):
+            PurchaseOrder.push_to_report_cashflow(purchase_order)
     print('reset_and_run_reports_sale done.')
